feat: adjust proplancontrols to only account for pro changes
