--- src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/ProPlanControls/TotalPriceCallout/ProPrice/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ProPrice'
