--- src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/ProPlanControls/TotalPriceCallout/TeamPrice/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './TeamPrice'
