+++ src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/ProPlanControls/TotalPriceCallout/TotalPriceCallout.spec.tsx
@@ -8,9 +8,9 @@ import { MemoryRouter, Route } from 'react-router-dom'
 
 import { Plans } from 'shared/utils/billing'
 
-import TotalBanner from './TotalPriceCallout'
+import TotalPriceCallout from './TotalPriceCallout'
 
-const allPlans = [
+const availablePlans = [
   {
     marketingName: 'Basic',
     value: 'users-basic',
@@ -21,11 +21,10 @@ const allPlans = [
       'Unlimited public repositories',
       'Unlimited private repositories',
     ],
-    quantity: 1,
     monthlyUploadLimit: 250,
   },
   {
-    marketingName: 'Pro Team',
+    marketingName: 'Pro',
     value: 'users-pr-inappm',
     billingRate: 'monthly',
     baseUnitPrice: 12,
@@ -35,11 +34,10 @@ const allPlans = [
       'Unlimited private repositories',
       'Priority Support',
     ],
-    quantity: 10,
     monthlyUploadLimit: null,
   },
   {
-    marketingName: 'Pro Team',
+    marketingName: 'Pro',
     value: 'users-pr-inappy',
     billingRate: 'annually',
     baseUnitPrice: 10,
@@ -50,48 +48,21 @@ const allPlans = [
       'Priority Support',
     ],
     monthlyUploadLimit: null,
-    quantity: 10,
   },
   {
-    marketingName: 'Pro Team',
-    value: 'users-enterprisem',
+    marketingName: 'Team',
+    value: 'users-teamm',
     billingRate: 'monthly',
-    baseUnitPrice: 12,
-    benefits: [
-      'Configurable # of users',
-      'Unlimited public repositories',
-      'Unlimited private repositories',
-      'Priority Support',
-    ],
+    baseUnitPrice: 5,
+    benefits: ['Patch coverage analysis'],
     monthlyUploadLimit: null,
   },
   {
-    marketingName: 'Pro Team',
-    value: 'users-enterprisey',
-    billingRate: 'annually',
-    baseUnitPrice: 10,
-    benefits: [
-      'Configurable # of users',
-      'Unlimited public repositories',
-      'Unlimited private repositories',
-      'Priority Support',
-    ],
-    monthlyUploadLimit: null,
-  },
-  {
-    marketingName: 'Sentry',
-    value: 'users-sentrym',
-    billingRate: null,
-    baseUnitPrice: 0,
-    benefits: ['Includes 5 seats', 'Unlimited public repositories'],
-    monthlyUploadLimit: null,
-  },
-  {
-    marketingName: 'Sentry',
-    value: 'users-sentryy',
-    billingRate: null,
-    baseUnitPrice: 10,
-    benefits: ['Includes 5 seats', 'Unlimited private repositories'],
+    marketingName: 'Team',
+    value: 'users-teamy',
+    billingRate: 'yearly',
+    baseUnitPrice: 4,
+    benefits: ['Patch coverage analysis'],
     monthlyUploadLimit: null,
   },
 ]
@@ -126,7 +97,7 @@ afterAll(() => {
   server.close()
 })
 
-describe('TotalBanner', () => {
+describe('TotalPriceCallout', () => {
   afterEach(() => jest.resetAllMocks())
 
   function setup() {
@@ -138,7 +109,7 @@ describe('TotalBanner', () => {
         res(
           ctx.status(200),
           ctx.data({
-            owner: { availablePlans: allPlans },
+            owner: { availablePlans },
           })
         )
       )
@@ -157,7 +128,9 @@ describe('TotalBanner', () => {
       it('displays per month price', async () => {
         const { mockSetValue } = setup()
 
-        render(<TotalBanner {...props} setValue={mockSetValue} />, { wrapper })
+        render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
+          wrapper,
+        })
 
         const perMonthPrice = await screen.findByText(/\$100.00/)
         expect(perMonthPrice).toBeInTheDocument()
@@ -166,7 +139,9 @@ describe('TotalBanner', () => {
       it('displays billed annually at price', async () => {
         const { mockSetValue } = setup()
 
-        render(<TotalBanner {...props} setValue={mockSetValue} />, { wrapper })
+        render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
+          wrapper,
+        })
 
         const annualPrice = await screen.findByText(
           /\/per month billed annually at \$1,200.00/
@@ -177,7 +152,9 @@ describe('TotalBanner', () => {
       it('displays how much the user saves', async () => {
         const { mockSetValue } = setup()
 
-        render(<TotalBanner {...props} setValue={mockSetValue} />, { wrapper })
+        render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
+          wrapper,
+        })
 
         const moneySaved = await screen.findByText(/\$240.00/)
         expect(moneySaved).toBeInTheDocument()
@@ -192,7 +169,9 @@ describe('TotalBanner', () => {
 
       it('displays the monthly price', async () => {
         const { mockSetValue } = setup()
-        render(<TotalBanner {...props} setValue={mockSetValue} />, { wrapper })
+        render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
+          wrapper,
+        })
 
         const monthlyPrice = await screen.findByText(/\$120.00/)
         expect(monthlyPrice).toBeInTheDocument()
@@ -200,7 +179,9 @@ describe('TotalBanner', () => {
 
       it('displays what the user could save with annual plan', async () => {
         const { mockSetValue } = setup()
-        render(<TotalBanner {...props} setValue={mockSetValue} />, { wrapper })
+        render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
+          wrapper,
+        })
 
         const savings = await screen.findByText(/\$240.00/)
         expect(savings).toBeInTheDocument()
@@ -209,7 +190,7 @@ describe('TotalBanner', () => {
       describe('user switches to annual plan', () => {
         it('calls mock set value with pro annual plan', async () => {
           const { mockSetValue, user } = setup()
-          render(<TotalBanner {...props} setValue={mockSetValue} />, {
+          render(<TotalPriceCallout {...props} setValue={mockSetValue} />, {
             wrapper,
           })
