+++ src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/TeamPlanControls/PriceCallout/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './PriceCallout'
